+++ src/operationpoint/find_valid_initial_condition.jl
@@ -1,6 +1,6 @@
 using OrdinaryDiffEq: ODEFunction
 using NLsolve: nlsolve, converged
-
+using LinearAlgebra: pinv, UniformScaling
 
 struct RootRhs_ic
     rhs
@@ -13,8 +13,13 @@ function (rr::RootRhs_ic)(x)
 end
 
 function RootRhs_ic(of::ODEFunction)
-    mm = of.mass_matrix
-    @assert mm != nothing
+    if of.mass_matrix isa UniformScaling
+        n = length(of.syms)
+        mm = Array(of.mass_matrix, n, n)
+    else
+        mm = of.mass_matrix
+    end
+    @assert mm !== nothing
     mpm = pinv(mm) * mm
     RootRhs_ic(of.f, mpm)
 end