+++ test/runtests.jl
@@ -2,7 +2,11 @@ using Test
 using Crayons
 
 testlist = [
-    ("nodes/PQAlgebraic.jl", "PQAlgebraic Node Tests"),
+    ("common/States.jl", "States Tests"),
+    ("common/PowerGrid.jl", "PowerGrid Tests"),
+    ("common/PowerGridSolutions.jl", "PowerGridSolutions Tests"),
+    ("nodes/PQAlgebraic.jl", "PQAlgebraic Node Tests"), 
+    ("nodes/VoltageDependentLoad.jl", "VoltageDependentLoad Node Tests"),
     ("nodes/PVAlgebraic.jl", "PVAlgebraic Node Tests"),
     ("nodes/SlackAlgebraic.jl", "SlackAlgebraic Node Tests"),
     ("nodes/SwingEq.jl", "SwingEq Node Tests"),
@@ -11,6 +15,7 @@ testlist = [
     ("nodes/VoltageSourceInverterMinimal.jl", "VoltageSourceInverterMinimal Node Tests"),
     ("nodes/VoltageSourceInverterVoltagePT1.jl", "VoltageSourceInverterVoltagePT1 Node Tests"),
     ("nodes/CurrentSourceInverterMinimal.jl", "CurrentSourceInverterMinimal Node Tests"),
+    ("nodes/experimental/CompositeNode.jl", "CompositeNode Node Tests"),
     ("nodes/ExponentialRecoveryLoad.jl", "ExponentialRecoveryLoad Node Tests"),
     ("nodes/FourthOrderEqGovernorExciterAVR.jl", "FourthOrderEqGovernorExciterAVR Node Tests"),
     ("nodes/experimental/RLCLoad.jl", "RLCLoad Node Tests"),
@@ -20,11 +25,13 @@ testlist = [
     ("nodes/experimental/CurtailedPowerPlantWithInertia.jl","CurtailedPowerPlantWithInertia Tests"),
     ("lines/StaticLine.jl", "StaticLine tests"),
     ("lines/PiModelLine.jl", "PiModelLine tests"),
-    ("common/PowerGrid.jl", "PowerGrid Tests"),
-    ("common/States.jl", "States Tests"),
-    ("simulations/PowerGridSolutions.jl", "PowerGridSolutions Tests"),
-    ("simulations/simulations.jl", "Simulation Tests"),
-    ("simulations/PowerPerturbation.jl", "PowerPerturbation Simulation Tests"),
+    ("lines/RLLine.jl", "RLLine tests"),
+    ("lines/Transformer.jl", "Transformer tests"),
+    ("faults/ChangeInitialConditions.jl", "ChangeInitialConditions Tests"),
+    ("faults/NodeParameterChange.jl", "NodeParameterChange Tests"),
+    ("faults/LineFailure.jl", "LineFailure Tests"),
+    ("faults/PowerPerturbation.jl", "PowerPerturbation Simulation Tests"),
+    ("faults/NodeShortCircuit.jl", "NodeShortCircuit Simulation Tests"),
     ("operationpoint/operationpoint.jl", "operationpoint Tests"),
     ("parsers/JsonParser.jl", "JsonParser Tests")
 ]